backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
SQLAlchemy>=2.0.35,<3.0
python-dotenv==1.0.1
Werkzeug==3.1.1
qrcode[pil]==7.4.2
Pillow==11.0.0
marshmallow==4.0.1